# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zhalwadi Samaj API Documentation",
        "description": "API documentation for Zhalwadi Samaj mobile application",
        "contact": {
            "email": "support@zhalwadisamaj.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://parivar.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/Kuldevi/get_all": {
            "get": {
                "tags": [
                    "Kuldevi"
                ],
                "summary": "Get All Kuldevi",
                "operationId": "2daa4fdbe8b472c2e77413e1092c587c",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/get_all": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get All Members",
                "description": "Retrieve a list of all members",
                "operationId": "d9644bfff4b801c18e5bc62535614251",
                "responses": {
                    "200": {
                        "description": "Members list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Members List"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/store": {
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Create New Member",
                "description": "Create a new member record",
                "operationId": "b13bdb994b0bc74c8bcb4bb2f89ff643",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "gu": {
                                                "type": "string",
                                                "example": "જ્હોન ડો"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "member@example.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member Created"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/get/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get Member by ID",
                "description": "Retrieve a specific member by ID",
                "operationId": "b36bd6349006c73dea59cec479848793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully get member."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your bearer token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}